requirements.txt
opencv-python>=4.5.0
pandas>=1.3.0
Pillow>=8.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
numpy>=1.21.0
easyocr>=1.4.0
matplotlib>=3.4.0
python-dotenv>=0.19.0
